+++ src/demo/Demo.vue
@@ -14,7 +14,7 @@ import { demos } from './utils/mocks.js'
     npm-install="@todovue/tv-sidebar"
     source-link="https://github.com/TODOvue/tv-sidebar"
     url-clone="https://github.com/TODOvue/tv-sidebar.git"
-    version="2.1.2"
+    version="2.2.0"
   />
 </template>